+++ tests/test_pawn.py
@@ -4,56 +4,139 @@ from game.pieces.pawn import Pawn
 
 class TestPawn(unittest.TestCase):
 
-    def test_str_white(self):
+    def setUp(self):
+        self.board = Board()  
+        for row in range(8):
+            for col in range(8):
+                self.board.set_piece(row, col, None)
+
+    def test_str(self):
         board = Board()
-        pawn = Pawn("WHITE", board)
+        pawn = Pawn("WHITE", self.board)
         self.assertEqual(
             str(pawn),
             "♟",
         )
 
-    def test_str_black(self):
-        board = Board()
-        pawn = Pawn("BLACK", board)
-        self.assertEqual(
-            str(pawn),
-            "♙",
-        )
+    # Blancos
 
     def test_move_white_initial(self):
         board = Board()
         pawn = Pawn("WHITE", board)
-        possibles = pawn.possible_positions(1, 0)
+        possibles = pawn.possible_movement_vertical_down(6, 0)
         self.assertEqual(
             possibles,
-            [(2, 0), (3, 0)]  # Assuming the pawn can move two steps forward on its first move
+            [(5, 0), (4, 0)]  
         )
+
+    def test_move_white_not_initial(self):
+        board = Board()
+        pawn = Pawn("WHITE", board)
+        possibles = pawn.possible_movement_vertical_down(4, 2)
+        self.assertEqual(
+            possibles,
+            [(3, 2)]  
+        )
+
+    def test_move_white_not_initial_blocked(self):
+        board = Board()
+        board.set_piece(3, 2, Pawn("WHITE", board))
+        pawn = Pawn("WHITE", board)
+        possibles = pawn.possible_movement_vertical_down(4, 2)
+        self.assertEqual(
+            possibles,
+            []  
+        )
+
+    def test_move_white_initial_blocked(self):
+        board = Board()
+        board.set_piece(4, 0, Pawn("WHITE", board))
+        pawn = Pawn("WHITE", board)
+        possibles = pawn.possible_movement_vertical_down(6, 0)
+        self.assertEqual(
+            possibles,
+            [(5, 0)]  
+        )
+
+    # def test_move_white_attack_right(self):
+    #     board = Board()
+    #     board.set_piece(5, 1, Pawn("BLACK", board))
+    #     pawn = Pawn("WHITE", board)
+    #     possibles = pawn.possible_attack_diagonal_down_right(6, 0)
+    #     self.assertEqual(
+    #         possibles,
+    #         [(5, 1)]  
+    #     )
+
+    # def test_move_white_attack_left(self):
+    #     board = Board()
+    #     board.set_piece(5, 1, Pawn("BLACK", board))
+    #     pawn = Pawn("WHITE", board)
+    #     possibles = pawn.possible_attack_diagonal_down_left(6, 2)
+    #     self.assertEqual(
+    #         possibles,
+    #         [(5, 1)]  
+    #     )
+
+    # Negros
 
     def test_move_black_initial(self):
         board = Board()
         pawn = Pawn("BLACK", board)
-        possibles = pawn.possible_positions(6, 0)
+        possibles = pawn.possible_movement_vertical_up(1, 0)
         self.assertEqual(
             possibles,
-            [(5, 0), (4, 0)]  # Assuming the pawn can move one step forward
+            [(2, 0), (3, 0)]  
         )
 
-    def test_move_white_not_inital(self):
-        board = Board()
-        pawn = Pawn("WHITE", board)
-        possibles = pawn.possible_positions(4, 2)
-        self.assertEqual(
-            possibles,
-            [(5, 2)]  # Assuming the pawn can move one step forward
-        )
-
-    def test_move_black_not_inital(self):
+    def test_move_black_not_initial(self):
         board = Board()
         pawn = Pawn("BLACK", board)
-        possibles = pawn.possible_positions(4, 2)
+        possibles = pawn.possible_movement_vertical_up(2, 2)
         self.assertEqual(
             possibles,
-            [(3, 2)]  # Assuming the pawn can move one step forward
+            [(3, 2)]  
         )
+
+    def test_move_black_not_initial_blocked(self):
+        board = Board()
+        board.set_piece(3, 2, Pawn("BLACK", board))
+        pawn = Pawn("BLACK", board)
+        possibles = pawn.possible_movement_vertical_up(2, 2)
+        self.assertEqual(
+            possibles,
+            []  
+        )
+    
+    def test_move_black_initial_blocked(self):
+        board = Board()
+        board.set_piece(3, 0, Pawn("BLACK", board))
+        pawn = Pawn("BLACK", board)
+        possibles = pawn.possible_movement_vertical_up(1, 0)
+        self.assertEqual(
+            possibles,
+            [(2, 0)]  
+        )
+
+    # def test_move_black_attack_right(self):
+    #     board = Board()
+    #     board.set_piece(1, 1, Pawn("WHITE", board))
+    #     pawn = Pawn("BLACK", board)
+    #     possibles = pawn.possible_positions(2, 0)
+    #     self.assertEqual(
+    #         possibles,
+    #         [(1, 1)]  
+    #     )
+
+    # def test_move_black_attack_left(self):
+    #     board = Board()
+    #     board.set_piece(1, 6, Pawn("WHITE", board))
+    #     pawn = Pawn("BLACK", board)
+    #     possibles = pawn.possible_positions(2, 7)
+    #     self.assertEqual(
+    #         possibles,
+    #         [(1, 6)]  
+    #    )
+
 if __name__ == '__main__':
     unittest.main()